Bufgix/statistics: Statistics routes for user application fixed

+++ backend/app/application/dto/statistic.py
@@ -1,4 +1,5 @@
 from datetime import date
+from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -22,7 +23,7 @@ class AvarageDaysDTO(BaseModel):
     id: int
     name: str
     color: str
-    average_days: int
+    average_days: Decimal
 
 
 class PlarformAppDTO(BaseModel):
